requirements.txt
atari_py==0.2.9
mindspore==2.0.0
numpy==1.21.6
numpy==1.21.5
opencv_contrib_python_headless==4.7.0.72
opencv_python==4.8.1.78
opencv_python_headless==4.7.0.72
PyYAML==6.0.1
